Organize field types
