+++ src/models/metadata-fields/field-parser-interface.ts
@@ -1,0 +1,3 @@
+export interface FieldParserInterface<T> {
+  parseValue(rawValue: string): T | undefined;
+}
