+++ test/models/metadata-fields/field-types/boolean.test.ts
@@ -1,0 +1,34 @@
+import { expect } from '@open-wc/testing';
+import { BooleanParser } from '../../../../src/models/metadata-fields/field-types/boolean';
+
+describe('BooleanParser', () => {
+  it('can parse string number truthy', async () => {
+    const parser = new BooleanParser();
+    const response = parser.parseValue('1');
+    expect(response).to.be.true;
+  });
+
+  it('can parse string number falsy', async () => {
+    const parser = new BooleanParser();
+    const response = parser.parseValue('0');
+    expect(response).to.be.false;
+  });
+
+  it('can parse words truthy', async () => {
+    const parser = new BooleanParser();
+    const response = parser.parseValue('true');
+    expect(response).to.be.true;
+  });
+
+  it('can parse words falsy', async () => {
+    const parser = new BooleanParser();
+    const response = parser.parseValue('false');
+    expect(response).to.be.false;
+  });
+
+  it('can parse date truthy', async () => {
+    const parser = new BooleanParser();
+    const response = parser.parseValue(Date());
+    expect(response).to.be.true;
+  });
+});
