+++ test/models/metadata-fields/field-types/date.test.ts
@@ -1,40 +1,5 @@
 import { expect } from '@open-wc/testing';
-
-import {
-  NumberParser,
-  DateParser,
-  DurationParser,
-  StringParser,
-  BooleanParser,
-} from '../../../src/models/metadata-fields/field-parsers';
-
-describe('NumberParser', () => {
-  it('can parse int strings', async () => {
-    const parser = new NumberParser();
-    const response = parser.parseValue('3');
-    expect(response).to.equal(3);
-  });
-
-  it('can parse float strings', async () => {
-    const parser = new NumberParser();
-    const response = parser.parseValue('3.14');
-    expect(response).to.equal(3.14);
-  });
-
-  it('returns undefined if the number is not a number', async () => {
-    const parser = new NumberParser();
-    const response = parser.parseValue('qab');
-    expect(response).to.be.undefined;
-  });
-});
-
-describe('StringParser', () => {
-  it('can parse strings', async () => {
-    const parser = new StringParser();
-    const response = parser.parseValue('3');
-    expect(response).to.equal('3');
-  });
-});
+import { DateParser } from '../../../../src/models/metadata-fields/field-types/date';
 
 describe('DateParser', () => {
   it('can parse date-only strings', async () => {
@@ -149,59 +114,5 @@ describe('DateParser', () => {
     const parser = new DateParser();
     const response = parser.parseValue('absjkdvfnskj');
     expect(response).to.be.undefined;
-  });
-});
-
-describe('DurationParser', () => {
-  it('can parse mm:ss format', async () => {
-    const parser = new DurationParser();
-    const response = parser.parseValue('45:23');
-    const expected = 23 + 45 * 60;
-    expect(response).to.equal(expected);
-  });
-
-  it('can parse hh:mm:ss format', async () => {
-    const parser = new DurationParser();
-    const response = parser.parseValue('3:45:23');
-    const expected = 23 + 45 * 60 + 3 * 60 * 60;
-    expect(response).to.equal(expected);
-  });
-
-  it('can parse decimal format', async () => {
-    const parser = new DurationParser();
-    const response = parser.parseValue('345.23');
-    expect(response).to.equal(345.23);
-  });
-});
-
-describe('BooleanParser', () => {
-  it('can parse string number truthy', async () => {
-    const parser = new BooleanParser();
-    const response = parser.parseValue('1');
-    expect(response).to.be.true;
-  });
-
-  it('can parse string number falsy', async () => {
-    const parser = new BooleanParser();
-    const response = parser.parseValue('0');
-    expect(response).to.be.false;
-  });
-
-  it('can parse words truthy', async () => {
-    const parser = new BooleanParser();
-    const response = parser.parseValue('true');
-    expect(response).to.be.true;
-  });
-
-  it('can parse words falsy', async () => {
-    const parser = new BooleanParser();
-    const response = parser.parseValue('false');
-    expect(response).to.be.false;
-  });
-
-  it('can parse date truthy', async () => {
-    const parser = new BooleanParser();
-    const response = parser.parseValue(Date());
-    expect(response).to.be.true;
   });
 });
